imprv: Can be deleted even on the grant pages

+++ apps/app/src/server/events/user.ts
@@ -27,7 +27,7 @@ class UserEvent extends EventEmitter {
 
     let page = await Page.findByPath(userHomepagePath, true);
 
-    if (page !== null && page.creator.toString() !== user._id.toString()) {
+    if (page != null && page.creator != null && page.creator.toString() !== user._id.toString()) {
       await this.crowi.pageService.deleteCompletelyUserHomeBySystem(userHomepagePath);
       page = null;
     }
